Updates asset paths and adds new image files: Reorganizes public asset paths by moving images to a dedicated directory for better structure and maintainability. Updates imports in components to reflect the new paths. Adds multiple new SVG and image files to the project.
