+++ components/home/LogoTicker.tsx
@@ -1,11 +1,11 @@
-import quantumLogo from "@/public/asset/quantum.svg";
-import acmeLogo from "@/public/asset/acme-corp.svg";
-import echoValleyLogo from "@/public/asset/echo-valley.svg";
-import pulseLogo from "@/public/asset/pulse.svg";
-import outsideLogo from "@/public/asset/outside.svg";
-import apexLogo from "@/public/asset/apex.svg";
-import celestialLogo from "@/public/asset/celestial.svg";
-import twiceLogo from "@/public/asset/twice.svg";
+import quantumLogo from "@/public/asset/images/quantum.svg";
+import acmeLogo from "@/public/asset/images/acme-corp.svg";
+import echoValleyLogo from "@/public/asset/images/echo-valley.svg";
+import pulseLogo from "@/public/asset/images/pulse.svg";
+import outsideLogo from "@/public/asset/images/outside.svg";
+import apexLogo from "@/public/asset/images/apex.svg";
+import celestialLogo from "@/public/asset/images/celestial.svg";
+import twiceLogo from "@/public/asset/images/twice.svg";
 import Image from "next/image";
 
 const logos = [
