+++ components/home/hero.tsx
@@ -4,8 +4,8 @@ import Balancer from "react-wrap-balancer"
 import { NavbarDemo } from "../navbar/navbar"
 import IntroBar from "@/app/components/introBar"
 import { TypingText } from "../ui/typing-text"
-import demoImage1 from "@/public/asset/1.jpg"
-import demoImage2 from "@/public/asset/2.png"
+import demoImage1 from "@/public/asset/images/1.jpg"
+import demoImage2 from "@/public/asset/images/2.png"
 import Waitlist from "./waitlist"
 import Image from "next/image"
 import Pointer from "./pointer"
